Moving tests for formatters into their own files
