+++ test/acceptance/cli-test.js
@@ -1,12 +1,10 @@
 import fs from 'node:fs';
-import path, { dirname } from 'node:path';
-import { fileURLToPath } from 'node:url';
+import path from 'node:path';
 
 import Project from '../helpers/fake-project.js';
 import run from '../helpers/run.js';
 import setupEnvVar from '../helpers/setup-env-var.js';
 
-const __dirname = dirname(fileURLToPath(import.meta.url));
 const ROOT = process.cwd();
 
 describe('ember-template-lint executable', function () {
@@ -623,66 +621,6 @@ describe('ember-template-lint executable', function () {
 
   describe('errors and warnings formatting', function () {
     describe('without --format=json param', function () {
-      it('should print properly formatted error messages', async function () {
-        project.setConfig({
-          rules: {
-            'no-bare-strings': true,
-          },
-        });
-        project.write({
-          app: {
-            templates: {
-              'application.hbs': '<h2>Here too!!</h2> <div>Bare strings are bad...</div>',
-              components: {
-                'foo.hbs': '{{fooData}}',
-              },
-            },
-          },
-        });
-
-        let result = await run(['.']);
-
-        expect(result.exitCode).toEqual(1);
-        expect(result.stdout.split('\n')).toEqual([
-          'app/templates/application.hbs',
-          '  1:4  error  Non-translated string used  no-bare-strings',
-          '  1:25  error  Non-translated string used  no-bare-strings',
-          '',
-          '✖ 2 problems (2 errors, 0 warnings)',
-        ]);
-        expect(result.stderr).toBeFalsy();
-      });
-
-      it('should print properly formatted error messages', async function () {
-        project.setConfig({
-          rules: {
-            'no-bare-strings': true,
-            'no-html-comments': true,
-          },
-        });
-        project.write({
-          app: {
-            templates: {
-              'application.hbs':
-                '<h2>Here too!!</h2><div>Bare strings are bad...</div><!-- bad html comment! -->',
-            },
-          },
-        });
-
-        let result = await run(['.']);
-
-        expect(result.exitCode).toEqual(1);
-        expect(result.stdout.split('\n')).toEqual([
-          'app/templates/application.hbs',
-          '  1:4  error  Non-translated string used  no-bare-strings',
-          '  1:24  error  Non-translated string used  no-bare-strings',
-          '  1:53  error  HTML comment detected  no-html-comments',
-          '',
-          '✖ 3 problems (3 errors, 0 warnings)',
-        ]);
-        expect(result.stderr).toBeFalsy();
-      });
-
       it('should be able run a rule passed in (rule:warn)', async function () {
         project.setConfig({
           rules: {
@@ -995,223 +933,6 @@ describe('ember-template-lint executable', function () {
 ✖ 3 problems (3 errors, 0 warnings)`
         );
 
-        expect(result.stderr).toBeFalsy();
-      });
-    });
-
-    describe('with --format=json param', function () {
-      it('should print valid JSON string with errors', async function () {
-        project.setConfig({
-          rules: {
-            'no-bare-strings': true,
-          },
-        });
-        project.write({
-          app: {
-            templates: {
-              'application.hbs': '<h2>Here too!!</h2> <div>Bare strings are bad...</div>',
-              components: {
-                'foo.hbs': '{{fooData}}',
-              },
-            },
-          },
-        });
-
-        let result = await run(['--format=json', '.']);
-
-        let expectedOutputData = {};
-        expectedOutputData['app/templates/application.hbs'] = [
-          {
-            column: 4,
-            line: 1,
-            endColumn: 14,
-            endLine: 1,
-            message: 'Non-translated string used',
-            filePath: 'app/templates/application.hbs',
-            rule: 'no-bare-strings',
-            severity: 2,
-            source: 'Here too!!',
-          },
-          {
-            column: 25,
-            line: 1,
-            endColumn: 48,
-            endLine: 1,
-            message: 'Non-translated string used',
-            filePath: 'app/templates/application.hbs',
-            rule: 'no-bare-strings',
-            severity: 2,
-            source: 'Bare strings are bad...',
-          },
-        ];
-
-        expect(result.exitCode).toEqual(1);
-        expect(JSON.parse(result.stdout)).toEqual(expectedOutputData);
-        expect(result.stderr).toBeFalsy();
-      });
-
-      it('should include information about fixing', async function () {
-        project.setConfig({
-          rules: {
-            'require-button-type': true,
-          },
-        });
-
-        project.write({
-          app: {
-            components: {
-              'click-me-button.hbs': '<button>Click me!</button>',
-            },
-          },
-        });
-
-        let result = await run(['.', '--format=json']);
-
-        let expectedOutputData = {};
-        expectedOutputData['app/components/click-me-button.hbs'] = [
-          {
-            column: 0,
-            line: 1,
-            endColumn: 26,
-            endLine: 1,
-            isFixable: true,
-            message: 'All `<button>` elements should have a valid `type` attribute',
-            filePath: 'app/components/click-me-button.hbs',
-            rule: 'require-button-type',
-            severity: 2,
-            source: '<button>Click me!</button>',
-          },
-        ];
-
-        expect(result.exitCode).toEqual(1);
-        expect(JSON.parse(result.stdout)).toEqual(expectedOutputData);
-        expect(result.stderr).toBeFalsy();
-      });
-    });
-
-    describe('with --format=json param and --quiet', function () {
-      it('should print valid JSON string with errors, omitting warnings', async function () {
-        project.setConfig({
-          rules: {
-            'no-bare-strings': true,
-            'no-html-comments': true,
-          },
-        });
-        project.write({
-          app: {
-            templates: {
-              'application.hbs':
-                '<h2>Here too!!</h2><div>Bare strings are bad...</div><!-- bad html comment! -->',
-            },
-          },
-        });
-
-        let result = await run(['.', '--format=json', '--quiet']);
-
-        let expectedOutputData = {};
-        expectedOutputData['app/templates/application.hbs'] = [
-          {
-            column: 4,
-            line: 1,
-            endColumn: 14,
-            endLine: 1,
-            message: 'Non-translated string used',
-            filePath: 'app/templates/application.hbs',
-            rule: 'no-bare-strings',
-            severity: 2,
-            source: 'Here too!!',
-          },
-          {
-            column: 24,
-            line: 1,
-            endColumn: 47,
-            endLine: 1,
-            message: 'Non-translated string used',
-            filePath: 'app/templates/application.hbs',
-            rule: 'no-bare-strings',
-            severity: 2,
-            source: 'Bare strings are bad...',
-          },
-          {
-            column: 53,
-            endColumn: 79,
-            endLine: 1,
-            filePath: 'app/templates/application.hbs',
-            fix: {
-              text: '{{! bad html comment! }}',
-            },
-            line: 1,
-            message: 'HTML comment detected',
-            rule: 'no-html-comments',
-            severity: 2,
-            source: '<!-- bad html comment! -->',
-          },
-        ];
-
-        expect(result.exitCode).toEqual(1);
-        expect(JSON.parse(result.stdout)).toEqual(expectedOutputData);
-        expect(result.stderr).toBeFalsy();
-      });
-
-      it('should exit without error and empty errors array', async function () {
-        project.setConfig({
-          rules: {
-            'no-html-comments': 'warn',
-          },
-        });
-        project.write({
-          app: {
-            templates: {
-              'application.hbs':
-                '<h2>Here too!!</h2><div>Bare strings are bad...</div><!-- bad html comment! -->',
-            },
-          },
-        });
-        let result = await run(['.', '--format=json', '--quiet']);
-
-        let expectedOutputData = {};
-        expectedOutputData['app/templates/application.hbs'] = [];
-
-        expect(result.exitCode).toEqual(0);
-        expect(JSON.parse(result.stdout)).toEqual(expectedOutputData);
-        expect(result.stderr).toBeFalsy();
-      });
-
-      it('should include information about fixing', async function () {
-        project.setConfig({
-          rules: {
-            'require-button-type': true,
-          },
-        });
-
-        project.write({
-          app: {
-            components: {
-              'click-me-button.hbs': '<button>Click me!</button>',
-            },
-          },
-        });
-
-        let result = await run(['.', '--format=json']);
-
-        let expectedOutputData = {};
-        expectedOutputData['app/components/click-me-button.hbs'] = [
-          {
-            column: 0,
-            line: 1,
-            endColumn: 26,
-            endLine: 1,
-            isFixable: true,
-            message: 'All `<button>` elements should have a valid `type` attribute',
-            filePath: 'app/components/click-me-button.hbs',
-            rule: 'require-button-type',
-            severity: 2,
-            source: '<button>Click me!</button>',
-          },
-        ];
-
-        expect(result.exitCode).toEqual(1);
-        expect(JSON.parse(result.stdout)).toEqual(expectedOutputData);
         expect(result.stderr).toBeFalsy();
       });
     });
@@ -1568,98 +1289,6 @@ describe('ember-template-lint executable', function () {
 
         expect(result.exitCode).toEqual(0);
         expect(fs.existsSync(path.join(project.baseDir, 'my-results.sarif'))).toEqual(true);
-      });
-
-      it('should be able to load relative formatter', async function () {
-        project.setConfig({
-          rules: {
-            'no-bare-strings': true,
-            'no-html-comments': true,
-          },
-        });
-        project.write({
-          app: {
-            templates: {
-              'application.hbs':
-                '<h2>Here too!!</h2><div>Bare strings are bad...</div><!-- bad html comment! -->',
-            },
-          },
-          'custom-formatter.js': `
-            class CustomFormatter {
-              constructor(options = {}) {
-                this.options = options;
-                this.console = options.console || console;
-              }
-
-              format(results) {
-                this.console.log(\`errors: \${results.errorCount}\`);
-                this.console.log(\`warnings: \${results.warningCount}\`);
-                this.console.log(\`fixable: \${(results.fixableErrorCount + results.fixableWarningCount)}\`);
-              }
-            }
-
-            module.exports = CustomFormatter;
-          `,
-        });
-
-        let result = await run(['.', '--format', './custom-formatter.js']);
-
-        expect(result.stdout).toMatchInlineSnapshot(`
-          "errors: 3
-          warnings: 0
-          fixable: 0"
-        `);
-        expect(result.stderr).toBeFalsy();
-      });
-
-      it('should be able to load formatter from node_modules', async function () {
-        project.setConfig({
-          rules: {
-            'no-bare-strings': true,
-            'no-html-comments': true,
-          },
-        });
-        project.write({
-          app: {
-            templates: {
-              'application.hbs':
-                '<h2>Here too!!</h2><div>Bare strings are bad...</div><!-- bad html comment! -->',
-            },
-          },
-        });
-
-        let fixturePath = path.resolve(
-          __dirname,
-          '..',
-          'fixtures',
-          'ember-template-lint-formatter-test'
-        );
-        let formatterDirPath = path.join(
-          project.baseDir,
-          'node_modules',
-          'ember-template-lint-formatter-test'
-        );
-
-        fs.mkdirSync(formatterDirPath);
-        fs.copyFileSync(
-          path.join(fixturePath, 'index.cjs'),
-          path.join(formatterDirPath, 'index.js')
-        );
-        fs.copyFileSync(
-          path.join(fixturePath, 'package.json'),
-          path.join(formatterDirPath, 'package.json')
-        );
-
-        let result = await run(['.', '--format', 'ember-template-lint-formatter-test']);
-
-        expect(result.stdout).toMatchInlineSnapshot(`
-          "Custom Formatter Header
-
-          errors: 3
-          warnings: 0
-          fixable: 0"
-        `);
-        expect(result.stderr).toBeFalsy();
       });
     });
